--- trendlyzer/trendlyzer-frontend/src/pages/Landing/App.js
@@ -1,20 +1,0 @@
-import { Login } from "./Login.tsx";
-import { Register } from "./Register.tsx";
-
-function App() {
-  const [currentForm, setCurrentForm] = useState('login');
-
-  const toggleForm = (formName) => {
-    setCurrentForm(formName);
-  }
-
-  return (
-    <div className="App">
-      {
-        currentForm === "login" ? <Login onFormSwitch={toggleForm} /> : <Register onFormSwitch={toggleForm} />
-      }
-    </div>
-  );
-}
-
-export default App;
